Move single-reference helper functions to readimage
